moved BackendForProxy to the proxy package
