+++ consensus/istanbul/proxy/types.go
@@ -25,6 +25,7 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/consensus"
 	"github.com/ethereum/go-ethereum/consensus/istanbul"
+	"github.com/ethereum/go-ethereum/p2p"
 	"github.com/ethereum/go-ethereum/p2p/enode"
 	"github.com/ethereum/go-ethereum/rlp"
 )
@@ -59,6 +60,67 @@ var (
 	// ErrNodeNotProxy is returned if this node is not a proxy
 	ErrNodeNotProxy = errors.New("node not a proxy")
 )
+
+// BackendForProxyEngine provides the Istanbul backend application specific functions for Istanbul proxy engine
+type BackendForProxyEngine interface {
+	// Address returns the owner's address
+	Address() common.Address
+
+	// IsProxiedValidator returns true if this node is a proxied validator
+	IsProxiedValidator() bool
+
+	// IsProxy returns true if this node is a proxy
+	IsProxy() bool
+
+	// SelfNode returns the owner's node (if this is a proxy, it will return the external node)
+	SelfNode() *enode.Node
+
+	// Sign signs input data with the backend's private key
+	Sign([]byte) ([]byte, error)
+
+	// Multicast sends a message to it's connected nodes filtered on the 'addresses' parameter (where each address
+	// is associated with those node's signing key)
+	// If sendToSelf is set to true, then the function will send an event to self via a message event
+	Multicast(addresses []common.Address, payload []byte, ethMsgCode uint64, sendToSelf bool) error
+
+	// GetValEnodeTableEntries retrieves the entries in the valEnodeTable filtered on the "validators" parameter.
+	// If the parameter is nil, then no filter will be applied.
+	GetValEnodeTableEntries(validators []common.Address) (map[common.Address]*istanbul.AddressEntry, error)
+
+	// RewriteValEnodeTableEntries will rewrite the val enode table with "entries" rows.
+	RewriteValEnodeTableEntries(entries []*istanbul.AddressEntry) error
+
+	// UpdateAnnounceVersion will notify the announce protocol that this validator's valEnodeTable entry has been updated
+	UpdateAnnounceVersion()
+
+	// SetEnodeCertificateMsg will set this node's enodeCertificate to be used for connection handshakes
+	SetEnodeCertificateMsgMap(enodeCertificateMsgMap map[enode.ID]*istanbul.Message) error
+
+	// RetrieveEnodeCertificateMsgMap will retrieve this node's handshake enodeCertificate
+	RetrieveEnodeCertificateMsgMap() map[enode.ID]*istanbul.Message
+
+	// VerifyPendingBlockValidatorSignature is a message validation function to verify that a message's sender is within the validator set
+	// of the current pending block and that the message's address field matches the message's signature's signer
+	VerifyPendingBlockValidatorSignature(data []byte, sig []byte) (common.Address, error)
+
+	// VerifyValidatorConnectionSetSignature is a message validation function to verify that a message's sender is within the
+	// validator connection set and that the message's address field matches the message's signature's signer
+	VerifyValidatorConnectionSetSignature(data []byte, sig []byte) (common.Address, error)
+
+	// RetrieveValidatorConnSet will retrieve the validator connection set.
+	// The parameter `retrieveCachedVersion` will specify if the function should retrieve the
+	// set directly from making an EVM call (which is relatively expensive), or from the cached
+	// version (which will be no more than one minute old).
+	RetrieveValidatorConnSet(retrieveCachedVersion bool) (map[common.Address]bool, error)
+
+	GenerateEnodeCertificateMsg(externalEnodeURL string) (*istanbul.Message, error)
+
+	// AddPeer will add a static peer
+	AddPeer(node *enode.Node, purpose p2p.PurposeFlag)
+
+	// RemovePeer will remove a static peer
+	RemovePeer(node *enode.Node, purpose p2p.PurposeFlag)
+}
 
 type ProxyEngine interface {
 	Start() error
